requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
opencv-python==4.10.0.84
easyocr==1.7.0
numpy
Pillow
python-dotenv
jinja2
